fixed rejection letter issues

+++ config/cron-tasks.js
@@ -21,21 +21,37 @@ module.exports = {
               },
             ],
           },
+          populate: {
+            appliedjob: {
+              populate: {
+                job: {
+                  title: true,
+                },
+              },
+              status: true,
+            },
+          },
         }
       );
 
       if (rejectionMails) {
         try {
           for (mail of rejectionMails) {
-            const res = await strapi.plugin("email").service("email").send({
-              to: mail.to,
-              from: "Koach Job <koachproject@gmail.com>",
-              subject: "Job Application",
-              text: mail?.body,
-            });
-
-            if (res.accepted) {
-              await strapi.entityService.delete("api::mail.mail", mail.id);
+            // Sen
+            if (mail?.appliedjob?.status === "REJECTED") {
+              console.log(mail);
+              const res = await strapi
+                .plugin("email")
+                .service("email")
+                .send({
+                  to: mail.to,
+                  from: "Koach Job <koachproject@gmail.com>",
+                  subject: `${mail?.appliedjob?.job?.title} Job Application`,
+                  text: mail?.body,
+                });
+              if (res.accepted) {
+                await strapi.entityService.delete("api::mail.mail", mail.id);
+              }
             }
           }
         } catch (error) {
@@ -44,6 +60,6 @@ module.exports = {
       }
     },
     // options: { rule: "0 0 * * 0" }, //This is run once every week
-    options: { rule: "* * * * *" }, //This is run once every 5minute
+    options: { rule: "* * * * *" },
   },
 };
